refactor(completion): extract import-map logic into SRP submodule (#0000)
